Refactor Connector into ConnectorRunner and Connector
ConnectorRunners implements the importing logic using a Connector
Connector implements the actual connector handling
Now, when new Connector-related operations which are not to be
performed in the scope of an import (e.g. list tables) they can be
added to the Connector class.
