+++ services/importer/lib/importer/exceptions.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 
 require_relative '../../../../services/datasources/lib/datasources/exceptions'
-require_relative './connector'
+require_relative './connector_runner'
 
 module CartoDB
   module Importer2
@@ -175,9 +175,9 @@ module CartoDB
       CartoDB::Datasources::DropboxPermissionError                => 1016,
       CartoDB::Datasources::BoxPermissionError                    => 1021,
       CartoDB::Datasources::GDriveNoExternalAppsAllowedError      => 1008,
-      CartoDB::Importer2::Connector::ConnectorError               => 1500,
-      CartoDB::Importer2::Connector::ConnectorsDisabledError      => 1501,
-      CartoDB::Importer2::Connector::InvalidParametersError       => 1502
+      Carto::Connector::ConnectorError               => 1500,
+      Carto::Connector::ConnectorsDisabledError      => 1501,
+      Carto::Connector::InvalidParametersError       => 1502
     }
   end
 end
